fix: add quote update methods and conversion handling in OpBase and CustomJson models
